[BE] Refactor: 보관함 앨범 서비스 분리 및 적용
